basic files structure
